requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.6.2
distro==1.9.0
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
openai==1.35.9
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
pydantic==2.8.0
pydantic_core==2.20.0
PyJWT==2.8.0
python-dotenv==1.0.1
sniffio==1.3.1
sqlparse==0.5.0
tqdm==4.66.4
typing_extensions==4.12.2
whitenoise==6.7.0

